requirements.txt
Django==4.2.20
psycopg2==2.9.11
Pillow==11.2.1
django-environ==0.11.2
whitenoise==6.9.0
openpyxl==3.1.5
reportlab==4.2.5
qrcode==8.1
